+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## 0.22.22
 
 - Do not list SDK packages in the outdated packages report.
+- Deprecated `downloadPackage`, use the new `ToolEnvironment.unpack` instead.
 - Fix issue with analyzing packages nested inside git repositories on windows.
   https://github.com/dart-lang/pana/issues/1207